+++ ml-kem/src/kem.rs
@@ -19,17 +19,6 @@ pub use ::kem::{Decapsulate, Encapsulate};
 /// A shared key resulting from an ML-KEM transaction
 pub(crate) type SharedKey = B32;
 
-#[cfg(all(feature = "pkcs8", feature = "alloc"))]
-use pkcs8::der::{Encode, asn1::BitStringRef};
-#[cfg(feature = "pkcs8")]
-use {
-    hybrid_array::Array,
-    pkcs8::{
-        der::{AnyRef, asn1::OctetStringRef},
-        spki::AssociatedAlgorithmIdentifier,
-    },
-};
-
 /// A `DecapsulationKey` provides the ability to generate a new key pair, and decapsulate an
 /// encapsulated shared key.
 #[derive(Clone, Debug)]
@@ -197,7 +186,7 @@ impl<P> EncapsulationKey<P>
 where
     P: KemParams,
 {
-    fn new(ek_pke: EncryptionKey<P>) -> Self {
+    pub(crate) fn new(ek_pke: EncryptionKey<P>) -> Self {
         let h = H(ek_pke.as_bytes());
         Self { ek_pke, h }
     }
@@ -287,144 +276,6 @@ where
         let dk = Self::DecapsulationKey::generate_deterministic(d, z);
         let ek = dk.encapsulation_key().clone();
         (dk, ek)
-    }
-}
-
-/// The serialization of the private key is a choice between three different formats
-/// [according to PKCS#8](https://lamps-wg.github.io/kyber-certificates/draft-ietf-lamps-kyber-certificates.html#name-private-key-format).
-///
-/// “For ML-KEM private keys, the privateKey field in `OneAsymmetricKey`
-/// contains one of the following DER-encoded `CHOICE` structures.
-/// The seed format is a fixed 64-byte `OCTET STRING` (66 bytes total
-/// with the 0x8040 tag and length) for all security levels,
-/// while the expandedKey and both formats vary in size by security level”
-#[cfg(feature = "pkcs8")]
-#[derive(Clone, Debug, pkcs8::der::Choice)]
-pub enum PrivateKeyChoice<'o> {
-    /// FIPS 203 format for an ML-KEM private key: a 64-octet seed
-    #[asn1(tag_mode = "IMPLICIT", context_specific = "0")]
-    Seed(OctetStringRef<'o>),
-}
-
-#[cfg(feature = "pkcs8")]
-impl<P> AssociatedAlgorithmIdentifier for EncapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    type Params = P::Params;
-
-    const ALGORITHM_IDENTIFIER: pkcs8::spki::AlgorithmIdentifier<Self::Params> =
-        P::ALGORITHM_IDENTIFIER;
-}
-
-#[cfg(all(feature = "pkcs8", feature = "alloc"))]
-impl<P> pkcs8::EncodePublicKey for EncapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    /// Serialize the given `EncapsulationKey` into DER format.
-    /// Returns a `Document` which wraps the DER document in case of success.
-    fn to_public_key_der(&self) -> pkcs8::spki::Result<pkcs8::Document> {
-        let public_key = self.as_bytes();
-        let subject_public_key = BitStringRef::new(0, &public_key)?;
-
-        pkcs8::SubjectPublicKeyInfo {
-            algorithm: P::ALGORITHM_IDENTIFIER,
-            subject_public_key,
-        }
-        .try_into()
-    }
-}
-
-#[cfg(feature = "pkcs8")]
-impl<P> TryFrom<pkcs8::SubjectPublicKeyInfoRef<'_>> for EncapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    type Error = pkcs8::spki::Error;
-
-    /// Deserialize the encapsulation key from DER format found in `spki.subject_public_key`.
-    /// Returns an `EncapsulationKey` containing `ek_{pke}` and `h` in case of success.
-    fn try_from(spki: pkcs8::SubjectPublicKeyInfoRef<'_>) -> Result<Self, Self::Error> {
-        if spki.algorithm.oid != P::ALGORITHM_IDENTIFIER.oid {
-            return Err(pkcs8::spki::Error::OidUnknown {
-                oid: P::ALGORITHM_IDENTIFIER.oid,
-            });
-        }
-
-        let bitstring_of_encapsulation_key = spki.subject_public_key;
-        let enc_key = match bitstring_of_encapsulation_key.as_bytes() {
-            Some(bytes) => {
-                let arr: Array<u8, EncapsulationKeySize<P>> = match bytes.try_into() {
-                    Ok(array) => array,
-                    Err(_) => return Err(pkcs8::spki::Error::KeyMalformed),
-                };
-                EncryptionKey::from_bytes(&arr)
-            }
-            None => return Err(pkcs8::spki::Error::KeyMalformed),
-        };
-
-        Ok(Self::new(enc_key))
-    }
-}
-
-#[cfg(feature = "pkcs8")]
-impl<P> AssociatedAlgorithmIdentifier for DecapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    type Params = P::Params;
-
-    const ALGORITHM_IDENTIFIER: pkcs8::spki::AlgorithmIdentifier<Self::Params> =
-        P::ALGORITHM_IDENTIFIER;
-}
-
-#[cfg(all(feature = "pkcs8", feature = "alloc"))]
-impl<P> pkcs8::EncodePrivateKey for DecapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    /// Serialize the given `DecapsulationKey` into DER format.
-    /// Returns a `SecretDocument` which wraps the DER document in case of success.
-    fn to_pkcs8_der(&self) -> pkcs8::Result<pkcs8::SecretDocument> {
-        let decaps_key_bytes = self.to_seed().ok_or(pkcs8::Error::KeyMalformed)?;
-        let pk_key_der =
-            PrivateKeyChoice::Seed(OctetStringRef::new(decaps_key_bytes.as_slice())?).to_der()?;
-        let pk_key_octetstr: OctetStringRef<'_> = OctetStringRef::new(&pk_key_der)?;
-
-        let private_key_info =
-            pkcs8::PrivateKeyInfoRef::new(P::ALGORITHM_IDENTIFIER, pk_key_octetstr);
-        pkcs8::SecretDocument::encode_msg(&private_key_info).map_err(pkcs8::Error::Asn1)
-    }
-}
-
-#[cfg(feature = "pkcs8")]
-impl<P> TryFrom<pkcs8::PrivateKeyInfoRef<'_>> for DecapsulationKey<P>
-where
-    P: KemParams + AssociatedAlgorithmIdentifier<Params = AnyRef<'static>>,
-{
-    type Error = pkcs8::Error;
-
-    /// Deserialize the decapsulation key from DER format found in `spki.private_key`.
-    /// Returns a `DecapsulationKey` containing `dk_{pke}`, `ek`, and `z` in case of success.
-    fn try_from(private_key_info_ref: pkcs8::PrivateKeyInfoRef<'_>) -> Result<Self, Self::Error> {
-        private_key_info_ref
-            .algorithm
-            .assert_algorithm_oid(P::ALGORITHM_IDENTIFIER.oid)?;
-
-        let decaps_key = match private_key_info_ref
-            .private_key
-            .decode_into::<PrivateKeyChoice>()
-        {
-            Ok(PrivateKeyChoice::Seed(seed)) => Self::from_seed(
-                seed.as_bytes()
-                    .try_into()
-                    .map_err(|_| pkcs8::Error::KeyMalformed)?,
-            ),
-            Err(_) => return Err(pkcs8::Error::KeyMalformed),
-        };
-
-        Ok(decaps_key)
     }
 }
